removed unnecessary casting and improved easter egg readability

+++ STTextBasedGame/GameHelpers.cs
@@ -11,7 +11,8 @@ namespace STTextBasedGame
     {
         public static void StrawberryDropper(Player player, int difficulty)
         {
-            int strawberries = Math.Max(1, 5 / (int)difficulty); // Drops fewer strawberries as difficulty increases
+            // If we ever implement an extra hard difficulty rememmber to add a Math.Ceiling to prevent the int from being rounded down
+            int strawberries = Math.Max(1, 5 / difficulty); // Drops fewer strawberries as difficulty increases
             player.Strawberry += strawberries;
             WriteColoredLine($"\nYou defeated the wolf and looted it for {strawberries} strawberries!", ConsoleColor.Green);
         }
